# shared/openapi/leyo-swimming-v1.yaml
openapi: 3.0.3
info:
  title: leyoSwimming API
  description: 游泳私教预约平台后端 API 契约（登录注册 + 套餐购买订单退款批次）
  version: 1.1.0
  contact:
    name: leyoSwimming Team
...
servers:
  - url: http://localhost:8080
    description: 本地开发环境
paths:
...
  /api/coach/list:
    post:
      tags:
        - Coach
      summary: 教练列表查询
      description: 游客可访问，支持分页。
      operationId: listCoaches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCoachListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/coach/detail:
    post:
      tags:
        - Coach
      summary: 教练详情查询
      description: 游客可访问。错误码：COACH_NOT_FOUND
      operationId: getCoachDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCoachDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/package/list:
    post:
      tags:
        - Package
      summary: 正价套餐列表查询
      description: 游客可浏览平台标准套餐列表。
      operationId: listPackages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/coach/package/list:
    post:
      tags:
        - Coach
      summary: 教练关联套餐列表查询
      description: 查询某教练可售的标准套餐与自定义课包配置。错误码：COACH_NOT_FOUND
      operationId: listCoachPackages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachPackageListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCoachPackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/package/detail:
    post:
      tags:
        - Package
      summary: 套餐详情查询
      description: 查询套餐详情及适用教练。错误码：PACKAGE_NOT_FOUND
      operationId: getPackageDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePackageDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/user/package/list:
    post:
      tags:
        - Package
      summary: 我的套餐列表
      description: 按状态分组查询当前学员的套餐。
      operationId: listUserPackages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPackageListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserPackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/coach/package/detail:
    post:
      tags:
        - Coach
      summary: 套餐详情（教练/学员共用）
      description: 查询某个用户套餐的详情与使用记录。错误码：FORBIDDEN, PACKAGE_NOT_FOUND
      operationId: getUserPackageDetail
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPackageDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserPackageDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/package/refund:
    post:
      tags:
        - Order
      summary: 申请退款
      description: 发起套餐退款申请。错误码：REFUND_IN_PROGRESS, PACKAGE_ALREADY_REFUNDED, PACKAGE_FROZEN, REFUND_NOT_SUPPORTED, REFUND_EXPIRED, PACKAGE_EXHAUSTED_NOT_REFUNDABLE
      operationId: applyPackageRefund
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageRefundRequest'
      responses:
        '200':
          description: 申请成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePackageRefundResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/coach/student/list:
    post:
      tags:
        - Coach
      summary: 教练查看学员列表
      description: 查询当前教练关联学员，支持 active/history 分组。错误码：AUTH_FORBIDDEN
      operationId: listCoachStudents
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachStudentListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCoachStudentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/coach/student/detail:
    post:
      tags:
        - Coach
      summary: 教练查看学员资料
      description: 查询指定学员的完整资料。错误码：NOT_ASSOCIATED_STUDENT
      operationId: getCoachStudentDetail
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachStudentDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCoachStudentDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/coach/student/update:
    post:
      tags:
        - Coach
      summary: 教练更新学员资料
      description: 更新教练视角的学员学习档案。错误码：NOT_ASSOCIATED_STUDENT, READONLY_USER_PROFILE, IDEMPOTENCY_DUPLICATE
      operationId: updateCoachStudent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachStudentUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/coach/student/package/list:
    post:
      tags:
        - Coach
      summary: 教练查看学员套餐列表
      description: 查询某学员在当前教练名下的套餐。错误码：NOT_ASSOCIATED_STUDENT
      operationId: listCoachStudentPackages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachStudentPackageListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCoachStudentPackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/package/list:
    post:
      tags:
        - Admin Package
      summary: 管理员查看用户套餐列表
      description: 后台管理用户套餐分页查询。错误码：ADMIN_PERMISSION_DENIED
      operationId: listAdminPackages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackageListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponsePackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/package/detail:
    post:
      tags:
        - Admin Package
      summary: 管理员查看用户套餐详情
      description: 查询单个用户套餐的完整信息。错误码：ADMIN_PERMISSION_DENIED, PACKAGE_NOT_FOUND
      operationId: getAdminPackageDetail
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackageDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponsePackageDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/package/freeze:
    post:
      tags:
        - Admin Package
      summary: 管理员冻结套餐
      description: 冻结用户套餐。错误码：ADMIN_PERMISSION_DENIED, PACKAGE_NOT_FOUND, PACKAGE_NOT_ACTIVE, PACKAGE_CONCURRENTLY_UPDATED
      operationId: freezeAdminPackage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackageFreezeRequest'
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponseMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/admin/package/unfreeze:
    post:
      tags:
        - Admin Package
      summary: 管理员解冻套餐
      description: 解冻用户套餐。错误码：ADMIN_PERMISSION_DENIED, PACKAGE_NOT_FOUND, PACKAGE_NOT_FROZEN, PACKAGE_CONCURRENTLY_UPDATED
      operationId: unfreezeAdminPackage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackageUnfreezeRequest'
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponseMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/admin/package/extend:
    post:
      tags:
        - Admin Package
      summary: 管理员延长套餐有效期
      description: 延长用户套餐有效期。错误码：ADMIN_PERMISSION_DENIED, PACKAGE_NOT_FOUND, INVALID_EXTENSION_REASON, PACKAGE_NOT_EXTENDABLE, PACKAGE_CONCURRENTLY_UPDATED
      operationId: extendAdminPackage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackageExtendRequest'
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponseMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/admin/package/refund:
    post:
      tags:
        - Admin Package
      summary: 管理员发起退款
      description: 管理员直接发起用户套餐退款。错误码：ADMIN_PERMISSION_DENIED, PACKAGE_NOT_FOUND, PACKAGE_STATUS_NOT_ALLOWED, PACKAGE_NOT_REFUNDABLE, REFUND_AMOUNT_INVALID, REFUND_PENDING_EXISTS, PACKAGE_CONCURRENTLY_UPDATED
      operationId: refundAdminPackage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackageRefundRequest'
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponseRefundOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
...
    PaginationRequest:
      type: object
      properties:
        page:
          type: integer
          default: 1
          minimum: 1
        pageSize:
          type: integer
          default: 10
          minimum: 1
          maximum: 50
    CoachListRequest:
      $ref: '#/components/schemas/PaginationRequest'
...
    CoachListPageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/CoachListItem'
    ApiResponseCoachListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CoachListPageResponse'
    CoachDetailRequest:
      type: object
      required:
        - coachId
      properties:
        coachId:
          type: integer
...
    CoachDetailResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gender:
          type: integer
        age:
          type: integer
        status:
          type: integer
        avatar:
          type: string
        rating:
          type: number
        yearsOfTeaching:
          type: integer
        totalStudents:
          type: integer
        totalHours:
          type: integer
        teachingStrokes:
          type: array
          items:
            type: string
        bio:
          type: string
        referencePrice:
          type: integer
        contact:
          $ref: '#/components/schemas/CoachDetailContact'
        certificates:
          type: array
          items:
            $ref: '#/components/schemas/CoachDetailCertificate'
        packages:
          type: array
          items:
            $ref: '#/components/schemas/CoachDetailPackage'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/CoachReview'
        availableTimes:
          type: array
          items:
            $ref: '#/components/schemas/AvailableTime'
        realTimeStatus:
          type: string
    ApiResponseCoachDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CoachDetailResponse'
...
    PackageListRequest:
      $ref: '#/components/schemas/PaginationRequest'
...
    PackageListPageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/PackageListItem'
    ApiResponsePackageListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PackageListPageResponse'
    CoachPackageListRequest:
      type: object
      required:
        - coachId
      properties:
        coachId:
          type: integer
...
    CoachPackageListResponse:
      type: object
      properties:
        coachId:
          type: integer
        coachStatus:
          type: integer
        referencePrice:
          type: integer
        standardPackages:
          type: array
          items:
            $ref: '#/components/schemas/CoachStandardPackage'
        customPackageEnabled:
          type: boolean
        customHoursMin:
          type: integer
        customHoursMax:
          type: integer
    ApiResponseCoachPackageListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CoachPackageListResponse'
    PackageDetailRequest:
      type: object
      required:
        - packageId
      properties:
        packageId:
          type: integer
        coachId:
          type: integer
...
    PackageDetailResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        packageMode:
          type: string
        hours:
          type: integer
        price:
          type: integer
        originalPrice:
          type: integer
        teachingType:
          type: string
        durationMinutes:
          type: integer
        validityDays:
          type: integer
        refundPolicySummary:
          type: string
        applicableCoaches:
          type: array
          items:
            $ref: '#/components/schemas/CoachBrief'
    ApiResponsePackageDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PackageDetailResponse'
...
    UserPackageListRequest:
      type: object
      properties:
        status:
          type: string
...
    UserPackageListResponse:
      type: object
      properties:
        active:
          $ref: '#/components/schemas/UserPackageGroup'
        exhausted:
          $ref: '#/components/schemas/UserPackageGroup'
        expired:
          $ref: '#/components/schemas/UserPackageGroup'
        refunded:
          $ref: '#/components/schemas/UserPackageGroup'
        frozen:
          $ref: '#/components/schemas/UserPackageGroup'
    ApiResponseUserPackageListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserPackageListResponse'
    UserPackageDetailRequest:
      type: object
      required:
        - packageId
      properties:
        packageId:
          type: integer
...
    UserPackageDetailResponse:
      type: object
      properties:
        packageId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        packageMode:
          type: string
        status:
          type: string
        totalHours:
          type: integer
        availableHours:
          type: integer
        reservedHours:
          type: integer
        consumedHours:
          type: integer
        expireAt:
          type: string
          format: date-time
        usageRecords:
          type: array
          items:
            $ref: '#/components/schemas/PackageUsageRecord'
    ApiResponseUserPackageDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserPackageDetailResponse'
...
    PackageRefundRequest:
      type: object
      required:
        - packageId
        - reasonType
      properties:
        packageId:
          type: integer
        reasonType:
          type: integer
          description: 1=教练原因，2=个人原因，3=平台原因
        reasonDetail:
          type: string
    PackageRefundResponse:
      type: object
      properties:
        refundOrderId:
          type: integer
        status:
          type: string
    ApiResponsePackageRefundResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PackageRefundResponse'
    CoachStudentListRequest:
      allOf:
        - $ref: '#/components/schemas/PaginationRequest'
        - type: object
          properties:
            tab:
              type: string
              enum: [active, history]
              default: active
...
    CoachStudentListPageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        students:
          type: array
          items:
            $ref: '#/components/schemas/CoachStudentListItem'
    ApiResponseCoachStudentListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CoachStudentListPageResponse'
    CoachStudentDetailRequest:
      type: object
      required:
        - studentId
      properties:
        studentId:
          type: integer
...
    CoachStudentDetailResponse:
      type: object
      properties:
        studentUserId:
          type: integer
        userProfile:
          $ref: '#/components/schemas/UserProfile'
        coachSlice:
          $ref: '#/components/schemas/CoachStudentSlice'
    ApiResponseCoachStudentDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CoachStudentDetailResponse'
    CoachStudentUpdateRequest:
      type: object
      required:
        - studentId
        - idempotencyKey
      properties:
        studentId:
          type: integer
        learningStrokes:
          type: array
          items:
            type: string
        swimLevel:
          type: string
        basics:
          type: string
        notes:
          type: string
        idempotencyKey:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ApiResponseMessageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/MessageResponse'
    CoachStudentPackageListRequest:
      allOf:
        - $ref: '#/components/schemas/PaginationRequest'
        - type: object
          required:
            - studentId
          properties:
            studentId:
              type: integer
...
    CoachStudentPackageListPageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        packages:
          type: array
          items:
            $ref: '#/components/schemas/CoachStudentPackageItem'
    ApiResponseCoachStudentPackageListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CoachStudentPackageListPageResponse'
    AdminApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
...
    AdminApiResponseMessageResponse:
      allOf:
        - $ref: '#/components/schemas/AdminApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/MessageResponse'
    AdminPackageListRequest:
      allOf:
        - $ref: '#/components/schemas/PaginationRequest'
        - type: object
          properties:
            status:
              type: string
            packageMode:
              type: string
            keyword:
              type: string
            expireAtStart:
              type: string
              format: date-time
            expireAtEnd:
              type: string
              format: date-time
...
    AdminApiResponsePackageListResponse:
      allOf:
        - $ref: '#/components/schemas/AdminApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminPackageListPageResponse'
    AdminPackageDetailRequest:
      type: object
      required:
        - packageId
      properties:
        packageId:
          type: integer
...
    AdminApiResponsePackageDetailResponse:
      allOf:
        - $ref: '#/components/schemas/AdminApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminPackageDetailResponse'
    AdminPackageFreezeRequest:
      type: object
      required:
        - packageId
        - reasonDetail
        - version
      properties:
        packageId:
          type: integer
        reasonDetail:
          type: string
          enum:
            - 投诉处理中
            - 异常订单
            - 司法冻结
        version:
          type: integer
    AdminPackageUnfreezeRequest:
      type: object
      required:
        - packageId
        - version
      properties:
        packageId:
          type: integer
        version:
          type: integer
    AdminPackageExtendRequest:
      type: object
      required:
        - packageId
        - newExpireAt
        - reason
        - version
      properties:
        packageId:
          type: integer
        newExpireAt:
          type: string
          format: date-time
        reason:
          type: string
        version:
          type: integer
    AdminPackageRefundRequest:
      type: object
      required:
        - packageId
        - reason
        - version
      properties:
        packageId:
          type: integer
        reason:
          type: string
        refundAmount:
          type: integer
        adjustReason:
          type: string
        version:
          type: integer
...
    AdminApiResponseRefundOrderResponse:
      allOf:
        - $ref: '#/components/schemas/AdminApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RefundOrderResponse'
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 100001
            message: 请求参数错误
            data: null
    Unauthorized:
      description: 未授权或 token 失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 200001
            message: 登录已过期，请重新登录
            data: null
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 100002
            message: 资源不存在
            data: null
    Forbidden:
      description: 权限禁止或账号被禁用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 300001
            message: 账号已被禁用，请联系超级管理员
            data: null
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 500001
            message: 已有待审核申请
            data: null